03-developer-journey/pyproject.toml
[project]
name = "developer-journey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore-starter-toolkit>=0.2.5",
    "boto3>=1.42.25",
    "botocore>=1.42.25",
    "duckduckgo-search>=8.1.1",
    "ipykernel>=7.1.0",
    "ipython>=9.9.0",
    "jupyter>=1.1.1",
    "langfuse>=3.11.2",
    "mcp>=1.25.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "strands-agents-tools>=0.2.19",
    "strands-agents[otel]>=1.21.0",
    "tabulate>=0.9.0",
]
